requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Blockchain
web3==6.11.3
eth-account==0.9.0
eth-utils==2.3.1
eth-abi==4.2.1
eth-keys==0.4.0

# HTTP Clients
aiohttp==3.9.1
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Async Support
asyncio-contextmanager==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
python-json-logger==2.0.7

# Optional: Advanced Features
sqlalchemy==2.0.23
